refactor: 컴포넌트 이름 파스칼 형식으로 변경
