+++ src/app/template.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import PageTransition from '@/components/page-transition';
+import PageTransition from '@/components/PageTransition';
 
 export default function Template({ children }: { children: ReactNode }) {
   return <PageTransition>{children}</PageTransition>;
